fix: divide query cause of pageable with fetch join

+++ server-api/src/main/java/org/dongguk/dscd/wooahan/api/article/dto/response/ReadArticleListDto.java
@@ -28,7 +28,7 @@ public record ReadArticleListDto (
             @JsonProperty("created_at")
             LocalDateTime createdAt,
             @JsonProperty("comment_cnt")
-            Integer commentCnt,
+            Long commentCnt,
             @JsonProperty("nickname")
             String nickname,
             @JsonProperty("creator_id")
